Upgrade from actions/create-release to ncipollo/release-action
The README for actions/create-release mentions that it's currently
unmaintained, and points to some alternatives, including
ncipollo/release-action, which we use in other projects already.

The API is different, so this means changing how the parameters
are specified for the action, but it seems like a fairly
straightforward match.

The bonus is that ncipollo/release-action allows attaching
artifacts directly, which means:

* The three steps with actions/download-artifact have to happen
  earlier;

* The three steps with actions/upload-release-asset (which is also
  an archived, unmaintained action) don't need to happen at all; and

* No other steps refer to the `create_release` step id, so it
  can be removed.